add a workflow invocation feedback tab
This does not implement the backend API POST route for actually submitting an email yet. That part is also left draft in the client.

This commit:
- Adds the new tab for reporting any issues with invocations.
- Finds and shows steps with failed jobs, and workflow invocation message(s) if they exist.
- Adds a reusable `EmailReportForm` component which shows the user's email and has a text field. It requires a `submit` method as prop which it waits for to show any response messages.
